requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
factory-boy==3.2.0
Faker==8.12.1
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
kombu==5.1.0
mccabe==0.6.1
packaging==21.0
pluggy==1.0.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytest-django-factories==0.2.3
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
